requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
numpy==2.0.0
scikit-learn==1.5.1
Pillow==10.3.0
uvicorn==0.30.1
gunicorn==22.0.0